requirements.txt
python-resize-image==1.1.11